SNAPGPU-53: Triangular interpolation, egm96 and srtm3 base to float, finished up computeSlavePixPos

+++ test_integration/egm96/include/egm96_test.cuh
@@ -8,7 +8,7 @@ namespace tests{
 struct EGM96data{
     int max_lats;
     int max_lons;
-    double* egm;
+    float* egm;
     int size;
 };
